Resolution de l'issue #97. Implementation de la copie et du deplacement des fichiers et dossiers entre differents volumes.

+++ usr/lib/webos/ftp.js
@@ -5,47 +5,58 @@
  */
 
 Webos.FTPFile = function WFTPFile(data, point) {
-	data.ftppath = data.path;
-	data.path = point.getWebosPath(data.ftppath);
-	data.realpath = 'ftp://'+point.get('data').user+':'+point.get('data').password+'@'+point.get('data').host+':'+point.get('data').port+'/'+data.ftppath;
-	
-	data.readable = true;
-	data.writable = true;
-	
 	this._mountPoint = point;
 	
 	Webos.File.call(this, data); //On appelle la classe parente
 };
 Webos.FTPFile.prototype = {
+	hydrate: function(data) {
+		var point = this.get('mountPoint');
+
+		if (data.path) {
+			data.ftppath = data.path;
+			data.path = point.getWebosPath(data.ftppath);
+			data.realpath = 'ftp://'+point.get('data').user+':'+point.get('data').password+'@'+point.get('data').host+':'+point.get('data').port+'/'+data.ftppath;
+		}
+		
+		data.basename = (data.basename == '.') ? null : data.basename;
+		
+		data.readable = (typeof data.readable == 'undefined') ? true : data.readable;
+		data.writable = (typeof data.writable == 'undefined') ? true : data.writable;
+
+		return Webos.File.prototype.hydrate.call(this, data);
+	},
 	mountPoint: function() {
 		return this._mountPoint;
 	},
 	load: function(callback) {
 		var that = this;
 		callback = Webos.Callback.toCallback(callback);
-		
-		Webos.FTPFile.load(this.get('path'), this.get('mountPoint'), function(file) {
-			var updatedData = {};
-			for (var key in file.data()) {
-				if (key == 'path') {
-					continue;
-				}
-				if (that.get(key) !== file.data()[key]) {
-					updatedData[key] = file.data()[key];
-				}
+
+		return new Webos.ServerCall({
+			'class': 'FTPController',
+			'method': 'getMetadata',
+			'arguments': {
+				loginData: this.get('mountPoint').get('data'),
+				file: this.get('ftppath')
 			}
-			that.hydrate(updatedData);
+		}).load([function(response) {
+			var data = response.getData();
+
+			that._updateData(data);
+
 			callback.success(that);
-			for (var key in updatedData) {
-				that.notify('update', { key: key, value: updatedData[key] });
-			}
-		});
+		}, callback.error]);
 	},
 	rename: function(newName, callback) {
 		var that = this;
 		callback = Webos.Callback.toCallback(callback);
+
+		if (!this.checkAuthorization('write', callback)) {
+			return false;
+		}
 		
-		new Webos.ServerCall({
+		return new Webos.ServerCall({
 			'class': 'FTPController',
 			method: 'rename',
 			arguments: {
@@ -53,32 +64,23 @@ Webos.FTPFile.prototype = {
 				file: this.get('ftppath'),
 				newName: newName
 			}
-		}).load([function() {
-			that.hydrate({
-				path: that.get('dirname')+'/'+newName
-			});
-			that.load([function() {
-				callback.success(that);
-			}, function(response) {
-				callback.error(response, that);
-			}]);
+		}).load([function(response) {
+			that._updateData(response.getData());
+			
+			callback.success(that);
 		}, function(response) {
 			callback.error(response, that);
 		}]);
 	},
-	move: function(dest, callback) {
-		var that = this;
-		callback = Webos.Callback.toCallback(callback);
-		dest = String(dest);
-		
-		//TODO: moving files support
-		this._unsupportedMethod(callback);
-	},
 	remove: function(callback) {
 		var that = this;
 		callback = Webos.Callback.toCallback(callback);
+
+		if (!this.checkAuthorization('write', callback)) {
+			return false;
+		}
 		
-		new Webos.ServerCall({
+		return new Webos.ServerCall({
 			'class': 'FTPController',
 			method: 'delete',
 			arguments: {
@@ -87,17 +89,85 @@ Webos.FTPFile.prototype = {
 			}
 		}).load([function() {
 			that._remove();
+
 			callback.success();
 		}, function(response) {
 			callback.error(response, that);
 		}]);
 	},
-	contents: function(callback) {
+	readAsText: function(callback) {
+		var that = this;
+		callback = Webos.Callback.toCallback(callback);
+
+		if (!this.checkAuthorization('read', callback)) {
+			return false;
+		}
+		
+		if (typeof this._contents != 'undefined') {
+			callback.success(this._contents);
+			return;
+		}
+
+		if (this.get('is_dir')) {
+			this._unsupportedMethod(callback);
+			return false;
+		} else {
+			return new Webos.ServerCall({
+				'class': 'FTPController',
+				'method': 'getFile',
+				'arguments': {
+					loginData: this.get('mountPoint').get('data'),
+					file: this.get('ftppath')
+				}
+			}).load([function(response) {
+				var contents = response.getData().contents;
+				that._contents = contents;
+				callback.success(contents);
+			}, callback.error]);
+		}
+	},
+	readAsBinary: function(callback) {
 		var that = this;
 		callback = Webos.Callback.toCallback(callback);
 		
+		if (!this.checkAuthorization('read', callback)) {
+			return false;
+		}
+		
 		if (this.get('is_dir')) {
-			new Webos.ServerCall({
+			this._unsupportedMethod(callback);
+			return false;
+		} else {
+			return new Webos.ServerCall({
+				'class': 'FTPController',
+				method: 'getFileAsBinary',
+				arguments: {
+					loginData: this.get('mountPoint').get('data'),
+					file: this.get('ftppath')
+				}
+			}).load(new Webos.Callback(function(response) {
+				var contents = response.getData().contents;
+				callback.success(contents);
+			}, function(response) {
+				callback.error(response);
+			}));
+		}
+	},
+	contents: function(callback) {
+		var that = this;
+		callback = Webos.Callback.toCallback(callback);
+
+		if (!this.checkAuthorization('read', callback)) {
+			return false;
+		}
+		
+		if (typeof this._contents != 'undefined') {
+			callback.success(this._contents);
+			return;
+		}
+		
+		if (this.get('is_dir')) {
+			return new Webos.ServerCall({
 				'class': 'FTPController',
 				'method': 'getFileList',
 				'arguments': {
@@ -107,40 +177,39 @@ Webos.FTPFile.prototype = {
 			}).load([function(response) {
 				var data = response.getData();
 				var list = [];
-				for(var i in data) {
-					var file = new Webos.FTPFile(data[i], that.get('mountPoint'));
-					if (Webos.File._cache[file.get('path')]) {
-						Webos.File._cache[file.get('path')].hydrate(file.data());
-						file = Webos.File._cache[file.get('path')];
+				for (var i in data) {
+					var ftpFile = new Webos.FTPFile(data[i], that.get('mountPoint'));
+					var file = Webos.File.get(ftpFile.get('path'));
+					if (Webos.isInstanceOf(file, Webos.FTPFile)) {
+						file._updateData(ftpFile.data());
 					} else {
-						Webos.File._cache[file.get('path')] = file;
+						file._updateData({
+							is_dir: ftpFile.get('is_dir')
+						});
 					}
 					list.push(file);
 				}
+				that._contents = list;
 				callback.success(list);
 			}, callback.error]);
 		} else {
-			new Webos.ServerCall({
-				'class': 'FTPController',
-				'method': 'getFile',
-				'arguments': {
-					loginData: this.get('mountPoint').get('data'),
-					file: this.get('ftppath')
-				}
-			}).load([function(response) {
-				callback.success(response.getData().contents);
-			}, callback.error]);
+			return this.readAsText(callback);
 		}
 	},
-	setContents: function(contents, callback) {
+	writeAsText: function(contents, callback) {
 		var that = this;
 		callback = Webos.Callback.toCallback(callback);
 		
 		if (this.get('is_dir')) {
-			callback.error();
+			this._unsupportedMethod(callback);
+			return false;
+		}
+
+		if (!this.checkAuthorization('write', callback)) {
+			return false;
 		}
 		
-		new Webos.ServerCall({
+		return new Webos.ServerCall({
 			'class': 'FTPController',
 			'method': 'putFile',
 			'arguments': {
@@ -149,8 +218,44 @@ Webos.FTPFile.prototype = {
 				contents: contents
 			}
 		}).load([function(response) {
+			that._contents = contents;
+
+			that.notify('updatecontents', { contents: contents });
+
+			that._updateData(response.getData());
+
 			callback.success();
 		}, callback.error]);
+	},
+	writeAsBinary: function(contents, callback) {
+		var that = this;
+		callback = Webos.Callback.toCallback(callback);
+		
+		if (this.get('is_dir')) {
+			this._unsupportedMethod(callback);
+			return false;
+		}
+
+		if (!this.checkAuthorization('write', callback)) {
+			return false;
+		}
+		
+		return new Webos.ServerCall({
+			'class': 'FTPController',
+			'method': 'putFileAsBinary',
+			'arguments': {
+				loginData: this.get('mountPoint').get('data'),
+				file: this.get('ftppath'),
+				contents: contents
+			}
+		}).load([function(response) {
+			that._updateData(response.getData());
+
+			callback.success();
+		}, callback.error]);
+	},
+	setContents: function(contents, callback) {
+		return this.writeAsText(contents, callback);
 	}
 };
 Webos.inherit(Webos.FTPFile, Webos.File); //Héritage de Webos.File
@@ -225,17 +330,10 @@ Webos.FTPFile.get = function(file, point, data) {
 
 Webos.FTPFile.load = function(path, point, callback) {
 	callback = Webos.Callback.toCallback(callback);
-	
-	new Webos.ServerCall({
-		'class': 'FTPController',
-		'method': 'getMetadata',
-		'arguments': {
-			loginData: point.get('data'),
-			file: point.getRelativePath(path)
-		}
-	}).load([function(response) {
-		var file = new Webos.FTPFile(response.getData(), point);
-		
+
+	var file = Webos.FTPFile.get(path, point);
+
+	return file.load([function() {
 		callback.success(file);
 	}, callback.error]);
 };
@@ -243,7 +341,7 @@ Webos.FTPFile.load = function(path, point, callback) {
 Webos.FTPFile.createFile = function(path, point, callback) {
 	callback = Webos.Callback.toCallback(callback);
 	
-	new Webos.ServerCall({
+	return new Webos.ServerCall({
 		'class': 'FTPController',
 		'method': 'createFile',
 		'arguments': {
@@ -251,7 +349,8 @@ Webos.FTPFile.createFile = function(path, point, callback) {
 			file: point.getRelativePath(path)
 		}
 	}).load([function(response) {
-		var file = new Webos.FTPFile(response.getData(), point);
+		var file = Webos.FTPFile.get(path, point);
+		file._updateData(response.getData());
 		callback.success(file);
 	}, callback.error]);
 };
@@ -259,7 +358,7 @@ Webos.FTPFile.createFile = function(path, point, callback) {
 Webos.FTPFile.createFolder = function(path, point, callback) {
 	callback = Webos.Callback.toCallback(callback);
 	
-	new Webos.ServerCall({
+	return new Webos.ServerCall({
 		'class': 'FTPController',
 		'method': 'createFolder',
 		'arguments': {
@@ -267,7 +366,8 @@ Webos.FTPFile.createFolder = function(path, point, callback) {
 			dir: point.getRelativePath(path)
 		}
 	}).load([function(response) {
-		var file = new Webos.FTPFile(response.getData(), point);
+		var file = Webos.FTPFile.get(path, point);
+		file._updateData(response.getData());
 		callback.success(file);
 	}, callback.error]);
 };
